Feat :sparkles: Atualização para otimização to tempo de carregamento, passou pela exclusão do módulo de blog.
