+++ angular.json
@@ -47,8 +47,8 @@
 							"budgets": [
 								{
 									"type": "initial",
-									"maximumWarning": "3.5MB",
-									"maximumError": "4MB"
+									"maximumWarning": "500kB",
+									"maximumError": "1MB"
 								},
 								{
 									"type": "anyComponentStyle",
