+++ src/app/pages/home/home.component.html
@@ -35,7 +35,7 @@
 	}
 </div>
 
-<app-section #sectionSobre>
+<!-- <app-section #sectionSobre>
 	@defer (on viewport(sectionSobre)) {
 		<div class="flex flex-col gap-8 sm:gap-16">
 			<div class="mx-auto flex w-full max-w-2xl flex-col items-center gap-8">
@@ -62,7 +62,7 @@
 			</div>
 		</div>
 	}
-</app-section>
+</app-section> -->
 
 <app-section #sectionExperiences>
 	@defer (on viewport(sectionExperiences)) {
