+++ src/app/shared/components/nav-bar/nav-bar.component.ts
@@ -19,7 +19,6 @@ export class NavBarComponent implements OnInit {
 
 	protected links = [
 		{ label: 'Home', path: '/', exact: true },
-		{ label: 'Blog', path: '/blog', exact: false },
 		{ label: 'Contato', path: '/contato', exact: true }
 	];
